Keep external Slice data alive during Iterator lifetime
The data in the byte strings for the start and stop keys of an Iterator
(if any) are pointed to by LevelDB Slice instances (start_slice and
stop_slice). Since Slice instances do not copy the actual data, this
means the byte strings should be kept alive as long as the Slices are
used, i.e. during the complete lifetime of the iterator. Before this
change the byte strings where deallocated at the end of the constructor,
while the slices were used afterwards (resulting in random data or
segfaults). To fix this, the iterator now stores the start and stop byte
strings on the instance to avoid deallocation. This also means that the
has_start and has_stop flags are no longer needed, so those are removed.