Raise TypeError when attempting ‘in’, ‘not in’, and __contains__
The ‘in’ operator has fallback behaviour for objects that do not
specify __contains__, see
https://docs.python.org/3/reference/expressions.html#membership-test-details

Various Plyvel classes have `.__iter__()` methods, so this means that
‘key in db’ actually iterates over the db, which yields ‘(key, value)’
tuples, none of which will be equal to plain ‘key’. For
comparison, (key, value) in db actually works… accidentally!

For non-matches (always in practice) such as an innocent looking key
lookup means the whole database is iterated (extremely slow for larger
databases), and will always return False in the end. This is
unexpected and bad behaviour.

Instead, detect this and raise TypeError with a helpful error message.

Closes #26.